Barre de Vie ET de Furry !!
c'est beau ! j'en pleur des larmes... ... des vrai.

+++ TMMFT Version n°2/Assets/Model 3D/Guns/Couteau/Attaque.cs
@@ -10,7 +10,7 @@ public class Attaque : MonoBehaviour {
     public float PortéeCouteau;
     public int DegatArmeCouteau;
 
-    public int furry;
+    public float furry;
     private int furryMax;
     public int PointDeFurryGagnéParCoup;
